[R-0.1.3] Created exception handling and switched to response entity responses from the APIs

+++ src/main/java/ro/dragos/service/RoomService.java
@@ -3,12 +3,15 @@ package ro.dragos.service;
 import org.springframework.stereotype.Service;
 import ro.dragos.dto.RoomDto;
 import ro.dragos.dto.SeatDto;
+import ro.dragos.exceptions.IdUsedException;
+import ro.dragos.exceptions.NotFoundException;
 import ro.dragos.mappers.RoomMapper;
 import ro.dragos.mappers.SeatMapper;
 import ro.dragos.model.Room;
 import ro.dragos.model.Seat;
 import ro.dragos.repository.RoomRepository;
 import ro.dragos.repository.SeatRepository;
+import ro.dragos.utils.StringConstants;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -35,21 +38,20 @@ public class RoomService {
 
     public boolean addRoom(RoomDto roomDto) {
         if (roomDto == null) {
-            throw new IllegalArgumentException("Room was null");
+            throw new IllegalArgumentException(StringConstants.ROOM_NULL);
         }
 
         Room room = roomMapper.toEntity(roomDto);
 
         Optional<Room> roomWithSameId = roomRepository.findById(room.getId());
         if (roomWithSameId.isPresent()) {
-            return false;
+            throw new IdUsedException(StringConstants.ROOM_ID_USED);
         }
 
         try {
             roomRepository.save(room);
             return true;
         } catch (Exception e) {
-            room.getSeats().forEach(seatRepository::delete);
             roomRepository.delete(room);
             throw e;
         }
@@ -57,15 +59,15 @@ public class RoomService {
 
     public boolean updateRoom(Long roomId, RoomDto roomDto) {
         if (roomId == null || roomDto == null) {
-            throw new IllegalArgumentException("RoomId or room were null");
+            throw new IllegalArgumentException(StringConstants.ROOM_NULL);
         }
-
+        roomDto.setId(roomId);
         Room room = roomMapper.toEntity(roomDto);
 
         Optional<Room> roomBeforeUpdate = roomRepository.findById(roomId);
 
-        if (roomBeforeUpdate.isEmpty() || !roomId.equals(roomDto.getId())) {
-            return false;
+        if (roomBeforeUpdate.isEmpty()) {
+            throw new NotFoundException(StringConstants.ROOM_ID_NOT_FOUND);
         }
 
         Room oldRoom = roomBeforeUpdate.get();
@@ -77,20 +79,21 @@ public class RoomService {
         } catch (Exception e) {
             roomRepository.delete(room);
             roomRepository.save(oldRoom);
-            return false;
+            throw e;
         }
     }
 
     public boolean deleteRoom(Long roomId) {
         if (roomId == null) {
-            throw new IllegalArgumentException("roomId was null");
+            throw new IllegalArgumentException(StringConstants.ROOM_NULL);
         }
-        try {
-            roomRepository.deleteById(roomId);
-            return true;
-        } catch (Exception e) {
-            return false;
+        Optional<Room> optionalRoom = roomRepository.findById(roomId);
+        if (optionalRoom.isEmpty()) {
+            throw new NotFoundException(StringConstants.ROOM_ID_NOT_FOUND);
         }
+
+        roomRepository.deleteById(roomId);
+        return true;
     }
